Fixes cast exception issue with FailureDetectorMBeanMetricFamilyCollector.

+++ common/src/main/java/com/zegelin/prometheus/cassandra/collector/FailureDetectorMBeanMetricFamilyCollector.java
@@ -59,7 +59,7 @@ public class FailureDetectorMBeanMetricFamilyCollector implements MBeanGroupMetr
             final Stream<NumericMetric> phiMetricsStream = endpointPhiValues.stream().map(d -> {
                 final Labels labels = new Labels(ImmutableMap.of("endpoint", ((String) d.get("Endpoint")).substring(1)));
 
-                return new NumericMetric(labels, (Float) d.get("PHI"));
+                return new NumericMetric(labels, ((Double) d.get("PHI")).floatValue());
             });
 
             setBuilder.add(new GaugeMetricFamily("cassandra_endpoint_phi", null, phiMetricsStream));
